fix: Поправил модальные окна в соответствии с чеклистом
