+++ src/components/modal-overlay/modal-overlay.js
@@ -1,45 +1,25 @@
 import PropTypes from 'prop-types';
-import { useEffect } from 'react';
-import ReactDOM from 'react-dom';
 import overlayStyles from './modal-overlay.module.css';
-import { modalRoot } from './../../utils/constants';
 
 const ModalOverlay = (props) => {
-  useEffect(() => {
-    window.addEventListener('keydown', handleEscPress);
-    return () => window.removeEventListener('keydown', handleEscPress);
-  }, []);
-
-  const handleEscPress = (e) => {
-    if (e.key === 'Escape') {
-      props.closeModal();
-    }
-  }
-
   const handleOverlayClick = (e) => {
     if (e.target === e.currentTarget) {
       props.closeModal();
     }
   }
 
-  return ReactDOM.createPortal(
-    (
+  return (
     <div
       className={overlayStyles.overlay}
       onClick={handleOverlayClick}
     >
       {props.children}
     </div>
-    ),
-    modalRoot
   )
 };
 
-
 ModalOverlay.propTypes = {
-  title: PropTypes.string,
-  closeModal: PropTypes.func
+  closeModal: PropTypes.func.isRequired
 }
-
 
 export default ModalOverlay;
